+++ moveit_setup_assistant/include/moveit/setup_assistant/navigation_widget.h
