+++ moveit_setup_assistant/include/moveit/setup_assistant/setup_screen_widget.h
