+++ moveit_setup_assistant/include/moveit/setup_assistant/start_screen_widget.h
@@ -50,7 +50,7 @@
 #include <moveit/setup_assistant/tools/moveit_config_data.h>  // common datastructure class
 #endif
 
-#include "setup_screen_widget.h"  // a base class for screens in the setup assistant
+#include <moveit/setup_assistant/setup_screen_widget.h>  // a base class for screens in the setup assistant
 
 namespace moveit_setup_assistant
 {
