+++ moveit_setup_assistant/plugins/end_effectors_widget/end_effectors_widget.cpp
@@ -36,7 +36,7 @@
 
 // SA
 #include "end_effectors_widget.h"
-#include <widgets/header_widget.h>
+#include <moveit/setup_assistant/header_widget.h>
 // Qt
 #include <QFormLayout>
 #include <QMessageBox>
