+++ moveit_setup_assistant/plugins/planning_groups_widget/planning_groups_widget.cpp
@@ -48,7 +48,7 @@
 // ******************************************************************************************
 
 #include "ros/ros.h"
-#include <widgets/header_widget.h>
+#include <moveit/setup_assistant/header_widget.h>
 #include "planning_groups_widget.h"
 // Qt
 #include <QApplication>
@@ -59,9 +59,9 @@
 #include <QMessageBox>
 #include <QTreeWidgetItem>
 
-#include <widgets/double_list_widget.h>  // for joints, links and subgroups pages
-#include "kinematic_chain_widget.h"      // for kinematic chain page
-#include "group_edit_widget.h"           // for group rename page
+#include <moveit/setup_assistant/double_list_widget.h>  // for joints, links and subgroups pages
+#include "kinematic_chain_widget.h"                     // for kinematic chain page
+#include "group_edit_widget.h"                          // for group rename page
 
 // Cycle checking
 #include <boost/graph/adjacency_list.hpp>
