+++ moveit_setup_assistant/plugins/planning_groups_widget/planning_groups_widget.h
@@ -48,7 +48,7 @@ class QTreeWidgetItem;
 #include <moveit/setup_assistant/tools/moveit_config_data.h>
 #endif
 
-#include <widgets/setup_screen_widget.h>  // a base class for screens in the setup assistant
+#include <moveit/setup_assistant/setup_screen_widget.h>  // a base class for screens in the setup assistant
 
 // Forward Declaration (outside of namespace for Qt)
 class PlanGroupType;
