+++ moveit_setup_assistant/plugins/ros_controllers_widget/ros_controllers_widget.cpp
@@ -35,7 +35,7 @@
 
 // SA
 #include "ros_controllers_widget.h"
-#include <widgets/header_widget.h>
+#include <moveit/setup_assistant/header_widget.h>
 #include <moveit_msgs/JointLimits.h>
 // Qt
 #include <QFormLayout>
