+++ moveit_setup_assistant/plugins/simulation_widget/simulation_widget.h
@@ -46,8 +46,8 @@
 #include <moveit/setup_assistant/tools/moveit_config_data.h>
 #endif
 
-#include <widgets/header_widget.h>
-#include <widgets/setup_screen_widget.h>  // a base class for screens in the setup assistant
+#include <moveit/setup_assistant/header_widget.h>
+#include <moveit/setup_assistant/setup_screen_widget.h>  // a base class for screens in the setup assistant
 
 namespace moveit_setup_assistant
 {
