+++ moveit_setup_assistant/src/tools/collision_matrix_model.cpp
@@ -34,7 +34,7 @@
 
 /* Author: Robert Haschke */
 
-#include "collision_matrix_model.h"
+#include <moveit/setup_assistant/tools/collision_matrix_model.h>
 #include <boost/unordered_map.hpp>
 #include <boost/assign.hpp>
 #include <QVector>
