+++ moveit_setup_assistant/src/tools/rotated_header_view.cpp
@@ -34,7 +34,7 @@
 
 /* Author: Robert Haschke */
 
-#include "rotated_header_view.h"
+#include <moveit/setup_assistant/tools/rotated_header_view.h>
 #include <QPainter>
 
 namespace moveit_setup_assistant
